change "in" evaluator to use walk and simplify compiler (#3667)
This commit changes the "in" logic to walk the input value so its
implementation is now fully complete.  We also removed the network
type as an "in" operand since it was awkward to combine
the semantics of a cidr match with a test for value containment
in a nested value.  To this end, we added a cidr_match() function,
which also walks its value but it doesn't match the network
value straight up.

We also simplified the compiler and removed old vestiges
of filter compilation by removing compileFilter().  The logic
to determine when a constant "in" can be used is simpler
and more general now as we call EvalAtCompileTime instead of
checking for dag.Literal.

We also simplified the logic for regexp searches where a closure
was doing a redundant check with the walk as the walk checks the
top-level value not just values below.