has_error function
Add has_error function which will recurse a zed value and return true if
the value or any of its leafs an error. has_error memoizes the types it
has seen so it won't have to recurse types more than once (unless the it
contains an error type in a union in which case it'll always have to
recurse).

Closes #3719