add zed README
Ths doc adds the zed command README and factors out the design
discussion as design notes in an unlinked markdown doc design.md.

This needs quite a bit more work to add detail, document all the
options, and make the presentation more uniform but we will get to
this in our next big chunk of work on the lake.