rework expression compilation to use kernel.Builder (#3795)
This commit adapts the compiler to use methods on kernel.Builder
for compiling expressions.  This will be used in a subsequent PR
to allow dataflow operators to appear inside of expressions.
We also moved CompileAssignments to package expr and renamed it
NewAssignments since it was just creating runtime objects and not
actually doing any compilation.