reorganize layout of docs directory for docusaurus
This commit reorganizes the content under ./docs to better
interface with docusaurus.  This is the first commit of several
where we are focusing on structure here.  We will continue to
improve the content in a subsequent PRs and add links from the
top level repo README to the public docs site instead of pointing
into the docs markdown files.