streamline Zed lake format spec as work in progress (#3805)
This commit simplifies the Zed lake format spec and
clarifies that it needs to be updated.  We also
added some philosophy and framing in the introduction.

The text we removed was really design notes regarding
pending work and we pasted that into issue #3804 as a
placeholder to create and prioritize relevant issues.