Disable markdown check for github link
The markdown check for a link to documentation on the github rest API
has recently been consistently returning 403 and therefore causing the
test to fail even though the link is working for an unauthenticated
browser client. Disable the check on the specific link for now.
The issue is being tracked here: https://github.com/github/feedback/discussions/14773