Fix panic on dropping deeply nested records
Fix a bug in drop where dropping records nested with a depth of three or
more would cause a duplicate column panic. The issue is caused when a
third item is appended to a slice and go expands the capacity from 2 to 4. With
available capacity in the prefix when a fourth item is appended go does
not allocate a new array and insteads returns the same slice with the
last value changed. The code meanwhile expected append to always return a new
array which resulted in the same slice getting added multiple times
causing the panic.

Closes #3882