python: rework zed.Client.query_raw() to handle any format
The zed.Client.query_raw() method requests ZJSON, deserializes the JSON
values in the response, and returns an iterator over the resulting
Python dicitonaries.  Add a headers parameter to enable requesting any
format supported by the API (via an Accept header), remove the
deserialization, and return a requests.Response instead of an iterator.