Fix unflatten bug (#3913)
Fix bug where unflatten was producing incorrect results for certain
nested records. The previous approach where the record structure and
bytes value was constructed in a single pass, made the code difficult to
reason about. This pr reformats unflatten so it is done in two
passes: The first pass constructs the shape of the nested record and the
second pass builds the type the value for the record.

Closes #3904