Optimized pool key predicates
This pr introduces new functionality that uses pool key predicates in the
pushdown to optimize both object and object seek selection. Both objects
and their seek offsets are selected by using augmented version of the
filter used in the original query. Optimized pool key lookups are limited in
the following ways:

- Will only work on the basic comparison operators "==", ">=", ">",
  "<=", "<"
- The right hand side must be a literal.
- Will not work on expressions nested under unary operators.

This pr also removes the from pool range syntax that was the former
approach to limiting the scope of a query.