cli/outputflags: always default to ZNG for -split (#3938)
Flags.Init sets Flags.Format to "zson" when -split is specified, -o is
not specified, and standard output is attached to a terminal.  This is
surprising because -split writes to files and ZNG is the default for
files under all other circumstances.  Make ZNG the default here too.

Unfortunately, I don't see a straightforward way to test this.