refactor compiler a little bit (#3941)
This commit refactors the compiler and lake so the compiler can import
the lake and do more sophisticated lake things as we integrate warehouse
style optimizations in the Zed lake.  This reverses the previous pattern
where op.DataAdaptor was used by the compiler to talk to the lake.

For now, op.DataAdaptor still exists but we will get rid of this in a
subsequent PR and localize the lakey-ness disposition into the compiler.