remove mock lake and zc REPL
This commit removes the mock lake that was used for testing in favor
of an approach that uses actual lakes for testing the DAG optimizer.
This will make work on the optimizer easier and more reliable as we
add more functionality like the vector engine and summary objects.

As part of this, we updated all the ztests that covered DAG optimizations
to create an actual lake with the pool key under test.

While we were here, we also removed the REPL from zed dev compile.