remove op.DataAdaptor to get ready for dynamic planners (#3946)
This commit removes the op.DataAdaptor and creates a new data.Source
object under the compiler that has direct knowledge of the
storage engine and lake (when compiling to a lake).  This will
allow the compiler to build planing objects directly in the runtime
without having to go through the adaptor interface that would otherwise
have become ever more complex.

We added zbuf.Filter.Pushdown() as a temporary means to break the
import cycle between lake and compiler but this will go away once
the compiler implements pushdown directly into the planner entities.