add vector form of data objects to lake/data
This commit creates the vector concept for cloud data objects and
adds a simple test to make sure the creation and deletion of suchs
objects works.

We also updated some of the docs and code to replace the row/column
terminology with "sequence/vector" terminology to better describe
Zed's model of heterogeneous sequences of deeply typed values compared
to the relationable model of tables, schemas, columns, and rows.
There is more work to do here in the report so we created issue ##3949
to track it.