add error checks to journal replay and simplify snapshots
This commit adds error checks to the logic that plays a log
to build a snapshot.  This change exposed some bugs that
led to the realization that the snapshot abstraction was
tracking deleted stuff when such tracking belongs in Patches
not Snaphots.  We removed the deleted objects from Snapshot
and updated all the surrounding logic to use Patches instead.

While we were at it, we changed the branch merge logic to use
a new method on Patch to compute a diff between patches and also
detect deletion conflicts.  This allowed us to remove the
OverlappingDeletes check that is better done in the Diff computation.