service: fix GET /events Content-Type header
With #3870, the value of the Content-Type response header for GET
/events changed from text/event-stream to the negotiated content type.
This route is intended to provided server-sent events, for which
text/event-stream is the correct type, so revert to that (by using an
http.ResponseWriter instead of a service.ResponseWriter for
eventStreamWriter.body).

Unfortunately, I don't see a straightforward way to test this right now.