rework ZNG writer options (#3967)
zio/zngio.WriterOpts contains a single field, LZ4BlockSize, that both
enables compression and specifies the frame threshold, making it
impossible to create a writer with no compression and a positive frame
threshold.  Replace it with separate Compress and FrameThresh fields,
and accordingly, replace the -znglz4blocksize flag to zed, zq, et al.
with new -zngcompress and -zngframethresh flags.