add support for marshaling mixed-type arrays
This commit fixes a known bug where mixed-type arrays would be
erroneously marshaled into zng.  The fix is to detect when the
array has multiple types.  This is tricky because the type of
each element is only determined upon a recursive descent which
also builds the value.  So, we opportunistically go this route
with a temp builder and track all the types encounted.  If there
is more than one type, we throw  out the improperly built array
and recode it as an array of unions.

This could be made more efficient if we implemented a funciton
to recursively compute the Zed type of each Go value without
building the value.

Closes #2575