fix inconsitencies with ZSON type names (#4011)
This commit fixes inconsistencies between ZSON type names in the
implementation and the spec and also clarifies the spec a bit.
Type names can now be valid string using quotes along with being
identifiers with mixed dots as before.  We removed "/" from the allowable
characters and require quotes now in this case.