add "zed vector" command (#4016)
This commit adds a zed command for adding and deleting vectors by
attaching ZST files to data objects.  The presence of vectors is
tracked in the pool's journal and available to the snapshot.

This commit does not utilize the vectors in the query dag or
any data scan as this will begin to be implemented in subsequent commits.