add (partial) support for marshaling mixed-type arrays and zed.Type (#4029)
This commit partially fixes a known bug where mixed-type arrays
are erroneously marshaled into zng.  The fix is to detect when the
array has multiple types.  This is tricky because the type of
each element is only determined upon a recursive descent which
also builds the value.  So, we opportunistically go this route
with a temp builder and track all the types encounted.  If there
is more than one type, we throw  out the improperly built array
and recode it as an array of unions.

This solution currently works when the mixed-type array elements
are Zed named types, allowing the decoder to allocate the proper
concrete type for each underlying mixed type.  We're leaving
issue #4012 open to handle the case where the concrete types
can be recursively determined without the Zed named types.

We also added some tests that are skipped to be addressed by #4012
and added a test to cover the case where an interface value
points to a zero-valued concrete object.

Finally, we added support for marshaling type values to/from
interface values of zed.Type.  This is needed by the upcoming
changes to the ZST metadata implementation.