fix zson formatter bug with redefined named types
The zson formatter did not properly track type name bindings.
The data structure was backwards: you need to keep track
of the binding from a name to a type so that when bindings
change in DFS order they can be tracked by their name.

This came up in unions where the named type could be redefined
in a child of a previous type and thus needed to be rebound
instead of referring erronously to the child type.