update ZST format to use marshaled metadata (#4035)
This commit updates the ZST format to use marshaling instead of
hand-coded zed values in the metadata section.  To do so, we created
a new set of Go metadata structures that are simply marshaled into
the metadata section of the ZST file.  We bumped the version number
to reflect the change in format.

We also cleaned up a bunch of things in the zst package.

In subsequent PRs, we will change the name of package zst/column
to zst/vector and update the ZST specification.