simplify zst to use io.ReaderAt instead of storage.Seeker
This commit simplifies the zst package to use io.ReaderAt
throughout and get rid of storage.Seeker.  We also noticed
that the check in zstio for a seekable input did not work
and instead failed on the first call to ReadAt (which gives
an error for stdin).  We fixed this and added a test to
make sure stdin fails properly.