zio/jsonio: fix bug reading leading empty string
The jsonio.Reader.Read incorrectly interprets a leading empty string
("") as a null string.  This happens because Reader.buf is nil on the
first call to Reader.handleToken and, when the first token is an empty
string, remains nil for the call to Reader.builder.pushPrimativeItem.
Fix by initializing Reader.buf with a non-nil value.