python: extract package version from Git (#4061)
The Python package in python/zed installs as version 0.0.0, which isn't
helpful for support purposes.  Use setuptools_scm to extract a
meaningful version from Git.

The version is available via pip and (with Python >= 3.8) via
"importlib.metadata.version('zed')".  It is not available as a module
property because the package contains just the single-file module in
zed.py, so there's currently nowhere for setuptools_scm to write a
_version.py file.