runtime/exec: fix concurrent map writes in newSeekFinder (#4062)
newSeekFinder returns a closure containing a reference to a comparison
function from expr.NewValueCompareFn.  The comparison function isn't
safe for concurrent use, but because a lake may read objects in
parallel, the closure may be called concurrently, and thus the
comparison function too.  This causes crashes due to concurrent map
writes.  Fix by moving the call to expr.NewValueCompareFn into the
closure.