move misplaced ztests to re-enable them (#4065)
#3300 moved zio/anyio/ztests/json-array.yaml to zio/anyio/array.yaml,
disabling it.  Move it back to zio/anyio/ztests under the old name and
update it to reflect the change in #3555 to the handling of top-level
arrays in the JSON reader.

#3824 moved compiler/ztests/is-proc-field.yaml and
compiler/ztests/is-proc.yaml to is-field.yaml and is.yaml, disabling
them.  Move them back into compiler/ztests but keep the new names.