allow pool name pattern in from operator
In the from operator, allow use of a pattern (glob or regexp) as a pool
name, as in "from x* | sample" and "from (pool /x/ => sample)".

This is implemented in the semantic analysis pass by querying the lake
for all pools, identifying those with names matching the pattern,
creating a synthetic dag.From containing one trunk per matching pool,
and then incorporating the synthetic dag.From into the natural dag.From.

The synthetic dag.From is necessary so

    from (
      pool a => count()
      pool b* => count()
    )

will compile to

    from (
      pool a => count()
      pass => from (
        pool b1
        pool b2
        ...
      ) => count()
    )

However it's often unnecessary. For example, "from *" compiles to

    from (
      pass =>
        from (
          pool a
          ...
        )
    )

instead of the simpler

    from (
      pool a
      pool b
      ...
    )

An optimization to remove it when possible is future work.