runtime/expr: return error in newArrayOrSetStep (#4067)
newArrayOrSetStep returns early if newStep fails but does not return the
newStep error.  This can result in a shaper with a bad copyOp step
(because the zero-value step is a copyOp with a fromIndex of 0), which
can lead step.build to return malformed zed.Values.  Fix by returning
the error.