allow 'cut a["b"]'
'cut a["b"]' produces an error in compiler/semantic.semAssignment
because that function wants to use a["b"] as the assignment LHS, for
which it requires a dag.This, but semExpr returns a dag.BinaryExpr for
a["b"].  Remove the path-length restriction in isIndexOfThis so semExpr
will return a dag.This.

This change triggers test failures that expose the lack of map support
in runtime/expr.DotExpr.Eval, so add that.