move duplicate field check in zed.Context.LookupTypeRecord (#4123)
duplicateField appears in CPU profiles when parsing large numbers of
similarly-shaped JSON objects.  We can avoid most of those calls by
calling duplicateField only when we create a new record type.  This is
safe because the check for an existing record type will fail if there's
a duplicate field.

With this change, Context.mu is now held during the duplicateField call.
That shouldn't be a problem since creating a new record type should
happen rarely.