implement complex nulls consistently in ZST
We never quite implements nulls for complex types in ZST.
They were implemented in an ad hoc way using the "presence"
concept inside of fields and inside of unions.

We noted that primitives encode null explicitly with their
ZNG representation but complex types don't have a way to do
that in ZST since they are representing the arrangement of
primitives in a hierarchy so the nulls must be explicitly
encoded for each complex type.

In this new approach, we simplify the concept with a generic
Nulls metadata type that wraps each complex type.  The NullsWriter
encodes nulls/values as run lengths and if there are not any
nulls at EOF, it simply omits itself from the metadata output.
The NullsReader wraps an underlying Reader and returns the
sequence of nulls and underlying values by decoding the run lengths.