service: honor Content-Type header for load
The load endpoint ignores the Content-Type request header and always
performs auto-detection.  Honor the header and perform auto-detection
only when the header is empty or unset.

To keep tests passing, change api/client.Connection.Load to delete the
Content-Type header, which is set to application/x-zng in the request
returned by Connection.NewRequest.