Convert scan planner to runtime ops
This commit refactors all the Go native logics that schedules
data scans as elements of the run-time DAG.  This will make it
easier to extend the scanning logic with new optimizations like
use of the vector cache and more sophisticated search logic.
It will also make it easier to distribute the DAG across a cluster
as the objects to be scanned can now be serialized over the network
as Zed without the need for a specialiead protocol.

These changes represent a first step at the refactoring where we
have retained as much as possible the pre-existing logic.  Once
this has been merged, we will refactor further in a subsequent PR
to further simplify the logic, e.g., with some simplifying changes
to how deletion and search-index lookups work.  Also, we will
introduce package dag components for the new runtime ops to
further simplify debugging and tests.  This will be important as
we add further complexity and optimiztion to the scanning and
pushdown logic.

These planned changes are documented in #4212.

Closes #4163