cli: cancel Flags.Init context on SIGPIPE
As described at https://pkg.go.dev/os/signal#hdr-SIGPIPE, a Go program
that has not expressed interest in SIGPIPE will exit (immediately) on a
write to a broken pipe on file descriptor 1 or 2.  This prevents the zed
and zq commands from removing temporary directories when either of those
file descriptors is connected to a pager that exits before reading to
EOF.  Fix this in cli.Flags.Init this by adding SIGPIPE to the list of
signal.NotifyContext signals.

Fixes #3782.