remove bytes allocation in zio.Reader.Read implementations
Most zio.Reader.Read implementations allocate new storage for the
returned zed.Value.Bytes on each call in spite of the fact that the
implementation retains ownership of that storage and can reuse it.
Remove these extra allocations by switching from zcode.(*Builder).Reset
to Truncate and removing an unnecessary Reset in zio/jsonio.

This speeds up the JSON and Zeek readers by about 2% and the Parquet
reader by about 5%.