Fix meta queries on empty pools
Fix bug with running meta commit queries on empty pools. Previously the
error message would state that the meta type for the commit did not
exist when it did. Change this so the query remains a meta commit query
and a query on an empty snapshot is run.

This pr also fixes a panic that occurred when running a partitions meta
query on an empty pool.

Closes #3029