simplify scans of snapshot metadata
This commit simplifies how snapshot metadata is scanned by embracing
the fact that snapshots are stored entirely in memory.  To this end,
we removed the channels that streamed metadata as the original intention
was that the snapshot would not necessarily fit entirely in memory.
However, we can go a long way toward huge Zed lakes with the
in-memory constraint and when we want to scale further, we can
work a design that is consistent with where the Zed lake architecture
lands in the future.

We also removed the unnecessary calls the zbuf.MultiScanner when there
was only one scanner and updated some functions that did not need
to be exported.