service: support loading Parquet via API
When loading data via POST /pool/{pool}/branch/{branch}, if the request
content type is application/x-parquet, copy the request body to a
temporary file for the benefit of the Parquet reader, which needs a
reader that implements io.ReaderAt and io.Seeker.

In the future, we should probably add a way to disable this or limit
file size to prevent an adversary from filling the file system.

Closes #4228.