rename and move TestTranslateNameConflictUnion
This is a test of (*Context).TranslateType, so rename it to incoporate
the method name and move it to context_test.go with the other Context
method tests.

Since the test uses package zson, moving it requires changing the
context_test.go package to zed_test to avoid an import cycle.