replace zio.WarningReader with service.warningsReader (#4303)
zio.WarningReader is used only in service.handleBranchLoad, for which
it's overgeneralized.  It also formats its wrapped zio.Reader with %s
and prepends the result to warnings regardless of whether the reader
implements fmt.Stringer, with ugly and unhelpful results in
handleBranchLoad since a reader from anyio.NewReaderWithOpts doesn't
implement fmt.Stringer.

Replace it with service.warningsReader, which is simpler and doesn't
prepend the %s-formatted zio.Reader.