pkg/jsonlexer: change New parameter to io.Reader (#4311)
This package uses a bufio.Reader in Lexer instead of an io.ByteScanner
because the former improves performance slightly (0.5% to 1%) in
zio/jsonio.Reader.  Accordingly, New takes a *bufio.Reader parameter.
To make New a little more general, change its parameter to an io.Reader
and wrap that with bufio.NewReader.  (This doesn't affect
zio/jsonio.Reader performance because bufio.NewReader returns its
argument if it's a bufio.Reader with large enough size.)