add meta-data slicer (#4352)
This commit changes the algorithm for forming partitions of non-overlapping
metadata objects for in-order scans.  This gets us closer to fine-grained
metadata operators so we can add vcache integration to the compiler.

A major implication of these changes is that commit-order scanning
when pool keys are of uniform value is no longer maintained.  This
isn't as bad as it sounds because this feature never worked anyway
at larger scale (when the size of the uniform-key data exceeds the
batch parallelism of the merge op the order guarantee can fail and
we did not have test coverage for this).  See issue #4351.

We also simplified things to use data.Objects without extent.Spans and
changed the logic to use the underlying From/To fields of the data.Object.
There is more work to do on this front.

We rearranged the metadata operators to take an explicit zctx so there
can be type separation between the metadata processing and downstream
operators, but no type separation for metaqueries where the downstream
user query operates on the metadata.