replace keyfilter with pruner and unhook search index logic
This commit replaces the keyfilter pattern with a pruner pattern
that is more reminiscent of OLAP and data lake architectures.
The logic for detecting whether we can prune based on pool keys
has moved from the DAG package to the compiler optimizer and
now sits adjacent to the logic that decides whether filter predicates
can be pushed into the scanning layer.

We also added a "tap" option to the meta-queries to ease test
and debug.  When the tap keyword is appeneded to a pool:objects
query, the downstream flowgraph is compiled as if it were a
normal query but the query runs as a metaquery generates the
objects after they have been pruned in accordance with any
pushdowns present in the downstream flowgraph.  The flowgraph
is not applied to the stream of objects of course.

We also improved the seekindex implementation where we scan
only the seekindex frames that survive the pruner rather than
the smallest single span that covers all such frames as was
previously implemented.