add -auth.audience flag to "zed serve" (#4356)
The Auth0 API audience used by "zed serve" (i.e., the audience value in
the GET /auth/method response that an authenticating client then sends
on to Auth0 in a device code request) is fixed at
https://lake.brimdata.io.  Replace that fixed value with an
-auth.audience flag to "zed serve".

The tenant ID and user ID claim names (in service/auth.TenantIDClaim and
UserIDClaim) still have a fixed prefix of https://lake.brimdata.io.  We
can replace that with the value of the -auth.audience flag separately.

For #4353.