unify data.Object and seekindex frames and improve seekindex
This commit unifies the data.Object and seekindex metadata so we
use the consistent pattern of From/To to represent a range of
Zed values.  This will simplify the pruning logic in a subsequent PR
where the pruning logic for data.Objects and seekindex frames can
be the same.

We now use the From/To pattern everywhere and no longer have First/Last
or Lower/Upper.

This commit also improves the seekindex data structure.  Instead of
having seekindex.Reader build the data structure that is desired
from the old seekindex format, we simply put the desired data structure
directly in the seekindex zng gile and use a standard zio.Reader
to unmarshal the index entries without additional processing.

We also got rid of data.Object.Meta and repositioned these fields
as top-level data.Object fields simplifying the layout here.

Because this changes the lake format, we bumped the lake version
from 1 to 2.