simplify lake version check and improve error message
This commit simplifies the lake version check and improves
the error message when running a new "zed serve" on an old lake.
Previously, the server would try to create a new lake when
the current lake didn't open and would report file exists
for the lake magic.  In this commit, we only try to create
a new lake if we get file not exists for the lake magic.

Closes #4387