convert from/to pattern to min/max for object metadata and seekindex
This commit changes the from/to pattern to an absolute min/max approach
for the range of values in a data.Object and seekindex.Entry.  Overall
this simplifies the code a bit with a small increae in complexity when
sorting object snapshots.  This will also make things consistent with
generalized pruning that we will implement in the vcache when we have
vector stats (including min/max).