single row implied aggregates default to values instead of records (#4420)
This commit changes the semantics of aggregate functions of an implied
summarize op to simply emit the values of the aggregation instead of
putting them in a record.  This changes the UX for aggregations to be
more Zed-like and less SQL-like when trying to form simple aggregations
that result in a single result without one or more group-by keys.

Many of us found ourselves typing the pattern 'agg() | yield agg' and
realized this approach would simplify things.  This way,
'echo "1 2 3"' | sum(this)' results in 6 instead of {sum:6}.

The change here is rather small but the ramifications on tests and docs are
non-trivial.