Allow fork operator with from and non-from legs
The fork operator requires that either all of its legs begin with a from
operator or that none do.  Relax this restriction to allow programs like
"fork (=> from (...) => pass)".

This is actually a side effect of the changes to
compiler/kernel.Builder.compileParallel that are necessary to implement
subquery join syntax, which requires a dag.Parallel node containing one
entry operation and one non-entry operation.

For #2839.