change "layout" pattern to "sort key"
This commit changes the naming convention for pool keys to use
the "sort key" instead of "layout" bringing this terminology more
in line with standard OLAP terms.  We kept the "layout" name in
the JSON config files of the lake so that we do not require a
new lake version.  We can change this name when we fix the sort
key to accommodate an order per key when using multiple keys.