move metadata ops into the DAG and refactor optimizer (#4526)
This commit makes a big shift in the optimizer to support the
forthcoming vector cache and engine.  With these changes, the
listing and manipulation of scanned objects is now part of the
DAG instead of being constructed at the last phase of runtime
generation.  This means we can implement DAG transormations
that replace or combine in vector operations with the traditional
sequence scanner.  This also will set us up for distributed
computation as the lister can transmit objects over the network
to different pieces of the distributed DAG.

We also adapted the "delete where" logic to use the refactored approach.

These changes improve performance when a scan is not order sensitive
as we now separate the slicing of objects into partitions from the
listing of raw objects.  SeqScanner and Deleter both handle reading
sorted and non-sorted cases by creating a merge-on-read scanner when
order is required for overlapping objects or simply scanning one object
at a time when order is not needed.