separate scope form sequential op
This commit creates a new operation for scopes and changes Sequential to
exclude local decls.  This gets us ready to remove the Sequential op altogether
and replace it was a new type Seq defined as []Op, which is not itself
an Op.  This is coming in a subsequent PR on branch "seq".

We also updated the grammar so that declarations require a parenthesized
scope block, except for the main scope, which requires no parentheses.