replace dag.Sequential with dag.Seq (as []dag.Op)
This commit replaces the sequential operator with a slice of ops
called dag.Seq imposing a more consistent and robust pattern on
the DAG.  This prevents us from playing games with sneaking
sequentials inside of sequentials as this was a questionable practice
that made downstream logic error prone.

This should make the DAG more robustness and easier to evolve
as we add support for vector ops.

To accommodate this approach, we added ast.Sample for the sample op
and put the logic to expand it into two ops in the semantic pass
where this belongs.

We also separated dag.Parallel into dag.Fork and dag.Scatter to get
rid of the Any flag inside of Parallel.