use json.Unmarshal pattern in unpack
This commit further simplies the unpacker by adopting the method
signature of json.Unmarshal.  Since reworking the packer design,
it became apparent that the unpackVal method could be used with
an arbitrary destination from the callee thereby providing the same
semantics of json.Unmarshal, which simplifies the logic.

This also created an opportunity to simplify the zjson decoder
to avoid the ugly workarounds with the previous unpack API.