Don't cancel context on SIGPIPE in "zed serve"
On Linux (but not macOS), "zed serve" exits if it writes to a closed
socket (i.e., a broken network connection) because that raises SIGPIPE,
canceling the context created by cmd/zed/serve.Command.Init due to the
change in #4180.  Fix this by removing SIGPIPE from the list of
signal.NotifyContext signals for "zed serve".

Closes #4498.