add dictionary encoding to vng
This commit adds dictionary encoding to the vng format where
primitive vectors are optionally encoded with dictionaries when
their cardinality is small.  Currently, the cardinality is limited
to 256 values; we will allow larger dictionaries in the future.

When the dictionary has length 1, the vector is simplified into
a "constant" vector and the dictionary and stats are omitted.